pyproject.toml
[project]
name = "mcf_rag_chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Casper Zanichell", email = "casperzanichelli98@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "lancedb[embeddings]>=0.27.0",
    "langchain-text-splitters>=1.1.0",
    "pyarrow>=23.0.0",
    "pydantic-ai>=1.44.0",
    "requests>=2.32.5",
    "streamlit>=1.53.0",
]

[project.scripts]
mcf_rag_chatbot = "mcf_rag_chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
